+++ Structs/Float3.cs
@@ -2,9 +2,10 @@
 {
     public struct Float3
     {
-        public float x;
-        public float y;
-        public float z;
+        public double x;
+        public double y;
+        public double z;
+
         public bool hasError;
 
         public override string ToString()
